added dropdown for user and refractor some code
