+++ client/src/components/Welcome.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import '../../../css/style.css'
+import '../../css/style.css'
 
 const Welcome = (props) => {
   return (
